requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
aiohttp
httpx
requests
aiosqlite
